Extract SectionMarker to shared component
Removes duplicated popover logic from all 4 panel files
and centralizes it in shared/SectionMarker.tsx
