+++ src/components/vignettes/multilingual/TranslationManagementPanel.tsx
@@ -5,6 +5,7 @@ import { motion, AnimatePresence } from 'framer-motion';
 import RichTextEditor from '@/components/demos/RichTextEditor';
 import { subtlePulse } from '@/lib/animations';
 import { multilingualContent } from '@/components/vignettes/multilingual/content';
+import { SectionMarker } from '@/components/vignettes/shared/SectionMarker';
 
 interface TranslationManagementPanelProps {
   className?: string;
@@ -15,66 +16,6 @@ interface TranslationManagementPanelProps {
 }
 
 type TranslationState = 'idle' | 'translating' | 'complete';
-
-interface SectionMarkerProps {
-  index: number;
-  noteId: string;
-  side: 'left' | 'right';
-  isActive: boolean;
-  onOpenChange: (noteId: string, isOpen: boolean) => void;
-  note: { label?: string; detail: string };
-}
-
-function SectionMarker({ index, noteId, side, isActive, onOpenChange, note }: SectionMarkerProps) {
-  const [open, setOpen] = useState(false);
-
-  const handleOpenChange = (isOpen: boolean) => {
-    setOpen(isOpen);
-    onOpenChange(noteId, isOpen);
-  };
-
-  // Dynamic import to avoid SSR issues
-  const Popover = require('@radix-ui/react-popover');
-
-  return (
-    <div
-      className={`absolute top-1/2 -translate-y-1/2 hidden lg:block ${
-        side === 'left' ? '-left-8' : '-right-8'
-      }`}
-    >
-      <Popover.Root open={open} onOpenChange={handleOpenChange}>
-        <Popover.Trigger asChild>
-          <button
-            className={`w-6 h-6 rounded-full text-xs font-semibold flex items-center justify-center
-                       bg-[var(--gold-500)] text-[var(--neutral-900)] hover:bg-[var(--gold-600)] transition-all cursor-pointer
-                       focus:outline-none focus-visible:ring-2 focus-visible:ring-[var(--gold-500)] focus-visible:ring-offset-2
-                       ${isActive ? 'ring-2 ring-[var(--gold-500)] ring-offset-2' : ''}`}
-            aria-label={`Design note ${index + 1}`}
-          >
-            {index + 1}
-          </button>
-        </Popover.Trigger>
-        <Popover.Portal>
-          <Popover.Content
-            side={side}
-            align="center"
-            sideOffset={12}
-            collisionPadding={16}
-            className="bg-white rounded-xl px-4 py-3 shadow-lg border border-gray-200 max-w-[280px] z-50"
-          >
-            {note.label && (
-              <p className="font-semibold text-sm text-gray-900">{note.label}</p>
-            )}
-            <p className={`text-sm text-gray-600 leading-relaxed ${note.label ? 'mt-1' : ''}`}>
-              {note.detail}
-            </p>
-            <Popover.Arrow className="fill-white" />
-          </Popover.Content>
-        </Popover.Portal>
-      </Popover.Root>
-    </div>
-  );
-}
 
 export default function TranslationManagementPanel({
   className = '',
